Rename Baseline substream modules
