+++ protocols/substreams/ethereum-baseline/src/modules/manual_updates.rs
