+++ protocols/substreams/ethereum-baseline/src/modules/slot_layout.rs
